various improvements to fix functionality and document
Co-authored with perplexity.ai: https://www.perplexity.ai/search/i-am-trying-to-deploy-a-google-BjRoJupjQ2eup440PQOEAQ
