+++ app.yaml
@@ -1,1 +1,1 @@
-runtime: nodejs10
+runtime: nodejs20
