+++ package.json
@@ -1,16 +1,16 @@
 {
-	"name": "gcp-proxy-func",
-	"version": "1.0.0",
-	"description": "Run http-proxy-middleware to relay requests",
-	"main": "index.js",
-	"scripts": {
-		"deploy": "gcloud functions deploy proxy --runtime nodejs8 --trigger-http --memory=128 --timeout=60s"
-	},
-	"dependencies": {
-		"cors": "^2.8.5",
-		"express": "^4.17.1",
-		"google-auth-library": "^7.0.4",
-		"googleapis": "^73.0.0",
-		"http-proxy-middleware": "^0.20.0"
-	}
+  "name": "gcp-proxy-func",
+  "version": "1.0.0",
+  "description": "Run http-proxy-middleware to relay requests",
+  "main": "index.js",
+  "scripts": {
+    "deploy": "gcloud functions deploy proxy --runtime nodejs20 --trigger-http --memory=128 --timeout=60s"
+  },
+  "dependencies": {
+    "cors": "^2.8.5",
+    "express": "^4.17.1",
+    "google-auth-library": "^7.0.4",
+    "googleapis": "^73.0.0",
+    "http-proxy-middleware": "^0.20.0"
+  }
 }
